reorganize loading operators and logic into a package
