+++ src/cc_catalog_airflow/dags/test_common_api_workflows.py
@@ -8,11 +8,13 @@ import util.config as conf
 SCRIPT = conf.SCRIPT
 CRONTAB_STR = conf.CRONTAB_STR
 RESOURCES = os.path.join(conf.API_SCRIPT_PATH, 'tests/resources')
+FILE_DIR = os.path.abspath(os.path.dirname(__file__))
 
 
-def test_dags_load_with_no_errors():
-    dag_bag = DagBag(include_examples=False)
-    dag_bag.process_file('common_api_workflows.py')
+def test_dags_load_with_no_errors(tmpdir):
+    tmp_directory = str(tmpdir)
+    dag_bag = DagBag(dag_folder=tmp_directory, include_examples=False)
+    dag_bag.process_file(os.path.join(FILE_DIR, 'common_api_workflows.py'))
     assert len(dag_bag.import_errors) == 0
